include text in the result report
Earlier the result screen didn't show the the text
Which was a bummer because you would have wanted to
read the text once more. For example in the case of
quotes.

That was not happening and I was not happy with it.
I made an executive decision and put all my money in
NFTees.

These are t-shirts that have images of NFTs that other
people have paid online for: <insert laughing guy meme 'ere>

Anyway, has any moose bit your sister's finger lately?