Change exit value to 0 from 1
When we are on the report page, and want to quit
Because the life has become so hard that you could
not carry on as it happens in case of maximus pain,
you just want to hit that ctrl+c and be done with it.

I tried in real life, doesn't work.
But in this program, we were exiting with 1.
That is just not good.
Had to return with 0.
That my unix-style guru said is the unix way.

But all my life I have known the unisex way.
Is there any other sex than mine?
Would one ever touch me who does not belong to
my sex?

I am also told by my guru that if I live in terminal,
probably not.